+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 * Added limits for "Columns" and "Rows" fields
 * FIX. Text color is not applied to all chiclets after selecting
 * FIX. Chiclets change place after switching between "Disabled" options
+