+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.4.5
+* Fixed issue: Multi selection doesn't work on mobile devices
+* Fixed issue: Incorrect default value for Chiclet Height and Width displayed
+* Fixed issue: Performance decrease with a lot of values
+
 ## 1.4.4
 * Fixed not matching lateral boundaries of search box and headers' outline
 * Fixed incorrect default value for Chiclet Height and Width displayed
@@ -7,3 +12,4 @@
 * Added limits for "Columns" and "Rows" fields
 * FIX. Text color is not applied to all chiclets after selecting
 * FIX. Chiclets change place after switching between "Disabled" options
+